Ensure that save_post() can run, even when $_POST['sticky-tax-term-id'] is not set.
Like with checkboxes, <select multiple> inputs will not be present in $_POST if no values are selected. This essentially makes it impossible to completely remove sticky terms from posts, which is obviously not cool.

Instead of checking that $_POST['sticky-tax-term-id] is set, assign an empty array to the key if it isn't already present, then let the rest of the script do its thing.