If a new taxonomy term is created for a non-hierarchical taxonomy, allow it to be added to the Sticky Tax meta box using its name instead of the term ID, prefixing it with the taxonomy name and a colon, then looking it up during save_post.
For example, assume someone is writing a post and has created the new Tag of "Coffee" — until the post is saved, this tag hasn't been added to the database, so there's no ID to pull. Instead, "Coffee" will be added with a value of "post_tag:Coffee", then save_post() will determine that the value is non-numeric, split it based on the colon (":"), and look up the term ID, since it should now be available.